requirements.txt
Django==1.11.6
django-extensions==1.9.6
djangorestframework==3.7.0
gunicorn==19.7.1
pep8==1.7.0
psycopg2==2.7.3.1
pytz==2017.2
six==1.11.0
